+++ resolwe_bio/utils/test.py
@@ -2,6 +2,7 @@
 import os
 import unittest
 import sys
+import time
 
 import wrapt
 
@@ -158,3 +159,8 @@ class KBBioProcessTestCase(BioProcessTestCase, LiveServerTestCase):
         super(KBBioProcessTestCase, cls).setUpClass()
         call_command('insert_features', os.path.join(TEST_FILES_DIR, 'features_gsea.tab.zip'))
         call_command('insert_mappings', os.path.join(TEST_FILES_DIR, 'mappings_gsea.tab.zip'))
+
+        # Wait for ElasticSearch to index the data.
+        # TODO: Better solution for ES5:
+        #       https://github.com/elastic/elasticsearch/pull/17986
+        time.sleep(2)